Reworked TopologyPanel constructor
