+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/topologyPane/TopologyPanel.java
@@ -19,10 +19,8 @@ import com.net2plan.gui.plugins.networkDesign.interfaces.ITopologyCanvas;
 import com.net2plan.gui.plugins.networkDesign.interfaces.ITopologyCanvasPlugin;
 import com.net2plan.gui.plugins.networkDesign.topologyPane.jung.AddLinkGraphPlugin;
 import com.net2plan.gui.plugins.networkDesign.topologyPane.jung.JUNGCanvas;
-import com.net2plan.gui.plugins.networkDesign.visualizationControl.VisualizationConstants;
 import com.net2plan.gui.plugins.networkDesign.visualizationControl.VisualizationState;
 import com.net2plan.gui.utils.FileDrop;
-import com.net2plan.gui.utils.JPopUpButton;
 import com.net2plan.interfaces.networkDesign.*;
 import com.net2plan.internal.Constants.DialogType;
 import com.net2plan.internal.Constants.NetworkElementType;
@@ -35,23 +33,24 @@ import org.apache.commons.collections15.BidiMap;
 import javax.swing.*;
 import javax.swing.border.LineBorder;
 import java.awt.*;
-import java.awt.event.*;
+import java.awt.event.ComponentAdapter;
+import java.awt.event.ComponentEvent;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseMotionAdapter;
 import java.io.File;
 import java.util.*;
 import java.util.List;
 
 @SuppressWarnings("unchecked")
-public class TopologyPanel extends JPanel implements ActionListener//FrequentisBackgroundPanel implements ActionListener//JPanel implements ActionListener
+public class TopologyPanel extends JPanel
 {
     private final GUINetworkDesign callback;
     private final ITopologyCanvas canvas;
 
-    private final JButton btn_increaseInterLayerDistance, btn_decreaseInterLayerDistance;
-    private final JButton btn_npChangeUndo, btn_npChangeRedo;
-    private final JToggleButton btn_showLowerLayerInfo, btn_showUpperLayerInfo, btn_showThisLayerInfo;
-    private final JLabel position;
     private final JPanel canvasPanel;
-    private final MultiLayerControlPanel multilayerControlPanel;
+
+    private final TopologyTopBar topBar;
+    private final TopologySideBar sideBar;
 
     private final File defaultDesignDirectory, defaultDemandDirectory;
 
@@ -93,140 +92,78 @@ public class TopologyPanel extends JPanel implements ActionListener//FrequentisB
      */
     public TopologyPanel(final GUINetworkDesign callback, File defaultDesignDirectory, File defaultDemandDirectory, Class<? extends ITopologyCanvas> canvasType, List<ITopologyCanvasPlugin> plugins)
     {
-        File currentDir = SystemUtils.getCurrentDir();
-
         this.callback = callback;
-        this.defaultDesignDirectory = defaultDesignDirectory == null ? new File(currentDir + SystemUtils.getDirectorySeparator() + "workspace" + SystemUtils.getDirectorySeparator() + "data" + SystemUtils.getDirectorySeparator() + "networkTopologies") : defaultDesignDirectory;
-        this.defaultDemandDirectory = defaultDemandDirectory == null ? new File(currentDir + SystemUtils.getDirectorySeparator() + "workspace" + SystemUtils.getDirectorySeparator() + "data" + SystemUtils.getDirectorySeparator() + "trafficMatrices") : defaultDemandDirectory;
-        this.multilayerControlPanel = new MultiLayerControlPanel(callback);
 
         try
         {
-            canvas = canvasType.getDeclaredConstructor(GUINetworkDesign.class, TopologyPanel.class).newInstance(callback, this);
-        } catch (Exception e)
-        {
-            throw new RuntimeException(e);
-        }
+            // File chooser default directory.
+            final File currentDir = SystemUtils.getCurrentDir();
 
-        if (plugins != null)
-            for (ITopologyCanvasPlugin plugin : plugins)
-                addPlugin(plugin);
+            this.defaultDesignDirectory = defaultDesignDirectory == null ? new File(currentDir + SystemUtils.getDirectorySeparator() + "workspace" + SystemUtils.getDirectorySeparator() + "data" + SystemUtils.getDirectorySeparator() + "networkTopologies") : defaultDesignDirectory;
+            this.defaultDemandDirectory = defaultDemandDirectory == null ? new File(currentDir + SystemUtils.getDirectorySeparator() + "workspace" + SystemUtils.getDirectorySeparator() + "data" + SystemUtils.getDirectorySeparator() + "trafficMatrices") : defaultDemandDirectory;
 
-        setLayout(new BorderLayout());
+            // Declare canvas : Reflections
+            this.canvas = canvasType.getDeclaredConstructor(GUINetworkDesign.class, TopologyPanel.class).newInstance(callback, this);
 
-        JPanel topPanel = new JPanel(new BorderLayout());
+            // Add given plugins
+            if (plugins != null)
+                for (ITopologyCanvasPlugin plugin : plugins)
+                    addPlugin(plugin);
 
-        TopologyToolBar toolbar = new TopologyToolBar(callback, this, canvas);
-        topPanel.add(toolbar, BorderLayout.NORTH);
+            this.setLayout(new BorderLayout());
 
-        add(topPanel, BorderLayout.NORTH);
+            // Top bar menu
+            this.topBar = new TopologyTopBar(callback, this, canvas);
+            this.add(topBar, BorderLayout.NORTH);
 
-        JComponent canvasComponent = canvas.getCanvasComponent();
+            // Canvas panel
+            final JComponent canvasComponent = canvas.getCanvasComponent();
+            canvasComponent.setBorder(LineBorder.createBlackLineBorder());
 
-        canvasPanel = new JPanel(new BorderLayout());
-        canvasComponent.setBorder(LineBorder.createBlackLineBorder());
+            this.canvasPanel = new JPanel(new BorderLayout());
 
-        JToolBar multiLayerToolbar = new JToolBar(JToolBar.VERTICAL);
-        multiLayerToolbar.setRollover(true);
-        multiLayerToolbar.setFloatable(false);
-        multiLayerToolbar.setOpaque(false);
+            this.sideBar = new TopologySideBar(callback, canvas);
 
-        canvasPanel.add(canvasComponent, BorderLayout.CENTER);
-        canvasPanel.add(multiLayerToolbar, BorderLayout.WEST);
-        add(canvasPanel, BorderLayout.CENTER);
+            this.canvasPanel.add(canvasComponent, BorderLayout.CENTER);
+            this.canvasPanel.add(sideBar, BorderLayout.WEST);
 
-        /* Multilayer buttons */
-        btn_increaseInterLayerDistance = new JButton();
-        btn_increaseInterLayerDistance.setToolTipText("Increase the distance between layers (when more than one layer is visible)");
-        btn_decreaseInterLayerDistance = new JButton();
-        btn_decreaseInterLayerDistance.setToolTipText("Decrease the distance between layers (when more than one layer is visible)");
-        btn_showLowerLayerInfo = new JToggleButton();
-        btn_showLowerLayerInfo.setToolTipText("Shows the links in lower layers that carry traffic of the picked element");
-        btn_showLowerLayerInfo.setSelected(callback.getVisualizationState().isShowInCanvasLowerLayerPropagation());
-        btn_showUpperLayerInfo = new JToggleButton();
-        btn_showUpperLayerInfo.setToolTipText("Shows the links in upper layers that carry traffic that appears in the picked element");
-        btn_showUpperLayerInfo.setSelected(callback.getVisualizationState().isShowInCanvasUpperLayerPropagation());
-        btn_showThisLayerInfo = new JToggleButton();
-        btn_showThisLayerInfo.setToolTipText("Shows the links in the same layer as the picked element, that carry traffic that appears in the picked element");
-        btn_showThisLayerInfo.setSelected(callback.getVisualizationState().isShowInCanvasThisLayerPropagation());
-        btn_npChangeUndo = new JButton();
-        btn_npChangeUndo.setToolTipText("Navigate back to the previous state of the network (last time the network design was changed)");
-        btn_npChangeRedo = new JButton();
-        btn_npChangeRedo.setToolTipText("Navigate forward to the next state of the network (when network design was changed");
+            this.add(canvasPanel, BorderLayout.CENTER);
 
-        // MultiLayer control window
-        JPopupMenu multiLayerPopUp = new JPopupMenu();
-        multiLayerPopUp.add(multilayerControlPanel);
-        JPopUpButton btn_multilayer = new JPopUpButton("", multiLayerPopUp);
+            // Buttons cannot be focusable
+            List<Component> children = SwingUtils.getAllComponents(this);
+            for (Component component : children)
+                if (component instanceof AbstractButton)
+                    component.setFocusable(false);
 
-        btn_increaseInterLayerDistance.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/increaseLayerDistance.png")));
-        btn_decreaseInterLayerDistance.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/decreaseLayerDistance.png")));
-        btn_multilayer.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/showLayerControl.png")));
-        btn_showThisLayerInfo.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/showLayerPropagation.png")));
-        btn_showUpperLayerInfo.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/showLayerUpperPropagation.png")));
-        btn_showLowerLayerInfo.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/showLayerLowerPropagation.png")));
-        btn_npChangeUndo.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/undoButton.png")));
-        btn_npChangeRedo.setIcon(new ImageIcon(TopologyPanel.class.getResource("/resources/gui/redoButton.png")));
+            // Action controllers
+            this.addPlugin(new PanGraphPlugin(callback, canvas, MouseEvent.BUTTON1_MASK)); // Panning
+            this.addPlugin(new PopupMenuPlugin(callback, this.canvas)); // Right button pop-up
 
-        btn_increaseInterLayerDistance.addActionListener(this);
-        btn_decreaseInterLayerDistance.addActionListener(this);
-        btn_showLowerLayerInfo.addActionListener(this);
-        btn_showUpperLayerInfo.addActionListener(this);
-        btn_showThisLayerInfo.addActionListener(this);
-        btn_npChangeUndo.addActionListener(this);
-        btn_npChangeRedo.addActionListener(this);
+            // Create links
+            if (callback.getVisualizationState().isNetPlanEditable() && this.getCanvas() instanceof JUNGCanvas)
+                addPlugin(new AddLinkGraphPlugin(callback, canvas, MouseEvent.BUTTON1_MASK, MouseEvent.BUTTON1_MASK | MouseEvent.SHIFT_MASK));
 
-        multiLayerToolbar.add(new JToolBar.Separator());
-        multiLayerToolbar.add(btn_multilayer);
-        multiLayerToolbar.add(btn_increaseInterLayerDistance);
-        multiLayerToolbar.add(btn_decreaseInterLayerDistance);
-        multiLayerToolbar.add(btn_showLowerLayerInfo);
-        multiLayerToolbar.add(btn_showUpperLayerInfo);
-        multiLayerToolbar.add(btn_showThisLayerInfo);
-        multiLayerToolbar.add(Box.createVerticalGlue());
-        //multiLayerToolbar.add(btn_npChangeUndo);
-        //multiLayerToolbar.add(btn_npChangeRedo);
+            // Move nodes
+            if (callback.getVisualizationState().isNetPlanEditable())
+                addPlugin(new MoveNodePlugin(callback, canvas, MouseEvent.BUTTON1_MASK | MouseEvent.CTRL_MASK));
 
-        this.addComponentListener(new ComponentAdapter()
-        {
-            @Override
-            public void componentResized(ComponentEvent e)
-            {
-                if (e.getComponent().getSize().getHeight() != 0 && e.getComponent().getSize().getWidth() != 0)
-                {
-                    canvas.zoomAll();
-                }
-            }
-        });
+            this.setBorder(BorderFactory.createTitledBorder(new LineBorder(Color.BLACK), "Network topology"));
 
-        List<Component> children = SwingUtils.getAllComponents(this);
-        for (Component component : children)
-            if (component instanceof AbstractButton)
-                component.setFocusable(false);
-
-        if (ErrorHandling.isDebugEnabled())
-        {
-            canvas.getCanvasComponent().addMouseMotionListener(new MouseMotionAdapter()
+            // Zoom all on resize
+            this.addComponentListener(new ComponentAdapter()
             {
                 @Override
-                public void mouseMoved(MouseEvent e)
+                public void componentResized(ComponentEvent e)
                 {
-                    Point point = e.getPoint();
-                    position.setText("view = " + point + ", NetPlan coord = " + canvas.getCanvasPointFromNetPlanPoint(point));
+                    if (e.getComponent().getSize().getHeight() != 0 && e.getComponent().getSize().getWidth() != 0)
+                    {
+                        canvas.zoomAll();
+                    }
                 }
             });
 
-            position = new JLabel();
-            add(position, BorderLayout.SOUTH);
-        } else
-        {
-            position = null;
-        }
-
-        new FileDrop(canvasComponent, new LineBorder(Color.BLACK), new FileDrop.Listener()
-        {
-            @Override
-            public void filesDropped(File[] files)
+            // File drop listener
+            new FileDrop(canvasComponent, new LineBorder(Color.BLACK), files ->
             {
                 for (File file : files)
                 {
@@ -240,69 +177,28 @@ public class TopologyPanel extends JPanel implements ActionListener//FrequentisB
                         break;
                     }
                 }
+            });
+
+            // DEBUG
+            if (ErrorHandling.isDebugEnabled())
+            {
+                final JLabel position = new JLabel();
+
+                canvas.getCanvasComponent().addMouseMotionListener(new MouseMotionAdapter()
+                {
+                    @Override
+                    public void mouseMoved(MouseEvent e)
+                    {
+                        Point point = e.getPoint();
+                        position.setText("view = " + point + ", NetPlan coord = " + canvas.getCanvasPointFromNetPlanPoint(point));
+                    }
+                });
+
+                add(position, BorderLayout.SOUTH);
             }
-        });
-
-        final ITopologyCanvasPlugin popupPlugin = new PopupMenuPlugin(callback, this.canvas);
-        addPlugin(new PanGraphPlugin(callback, canvas, MouseEvent.BUTTON1_MASK));
-        if (callback.getVisualizationState().isNetPlanEditable() && getCanvas() instanceof JUNGCanvas)
-            addPlugin(new AddLinkGraphPlugin(callback, canvas, MouseEvent.BUTTON1_MASK, MouseEvent.BUTTON1_MASK | MouseEvent.SHIFT_MASK));
-        addPlugin(popupPlugin);
-        if (callback.getVisualizationState().isNetPlanEditable())
-            addPlugin(new MoveNodePlugin(callback, canvas, MouseEvent.BUTTON1_MASK | MouseEvent.CTRL_MASK));
-
-        setBorder(BorderFactory.createTitledBorder(new LineBorder(Color.BLACK), "Network topology"));
-    }
-
-    @Override
-    public void actionPerformed(ActionEvent e)
-    {
-        Object src = e.getSource();
-        final VisualizationState vs = callback.getVisualizationState();
-        if (src == btn_increaseInterLayerDistance)
+        } catch (Exception e)
         {
-            if (vs.getCanvasNumberOfVisibleLayers() == 1) return;
-
-            final int currentInterLayerDistance = vs.getInterLayerSpaceInPixels();
-            final int newInterLayerDistance = currentInterLayerDistance + (int) Math.ceil(currentInterLayerDistance * (VisualizationConstants.SCALE_IN - 1));
-
-            vs.setInterLayerSpaceInPixels(newInterLayerDistance);
-            canvas.updateInterLayerDistanceInNpCoordinates(newInterLayerDistance);
-            canvas.updateAllVerticesXYPosition();
-            canvas.refresh();
-        } else if (src == btn_decreaseInterLayerDistance)
-        {
-            if (vs.getCanvasNumberOfVisibleLayers() == 1) return;
-
-            final int currentInterLayerDistance = vs.getInterLayerSpaceInPixels();
-            int newInterLayerDistance = currentInterLayerDistance - (int) Math.ceil(currentInterLayerDistance * (1 - VisualizationConstants.SCALE_OUT));
-
-            if (newInterLayerDistance <= 0)
-                newInterLayerDistance = 1;
-
-            vs.setInterLayerSpaceInPixels(newInterLayerDistance);
-            canvas.updateInterLayerDistanceInNpCoordinates(newInterLayerDistance);
-            canvas.updateAllVerticesXYPosition();
-
-            canvas.refresh();
-        } else if (src == btn_showLowerLayerInfo)
-        {
-            vs.setShowInCanvasLowerLayerPropagation(btn_showLowerLayerInfo.isSelected());
-            canvas.refresh();
-        } else if (src == btn_showUpperLayerInfo)
-        {
-            vs.setShowInCanvasUpperLayerPropagation(btn_showUpperLayerInfo.isSelected());
-            canvas.refresh();
-        } else if (src == btn_showThisLayerInfo)
-        {
-            vs.setShowInCanvasThisLayerPropagation(btn_showThisLayerInfo.isSelected());
-            canvas.refresh();
-        } else if (src == btn_npChangeUndo)
-        {
-            callback.requestUndoAction();
-        } else if (src == btn_npChangeRedo)
-        {
-            callback.requestRedoAction();
+            throw new RuntimeException(e);
         }
     }
 
@@ -485,17 +381,6 @@ public class TopologyPanel extends JPanel implements ActionListener//FrequentisB
     }
 
     /**
-     * Refreshes the name of a layer.
-     *
-     * @param layerId Layer identifier
-     * @since 0.3.1
-     */
-    public void refreshLayerName(long layerId)
-    {
-        multilayerControlPanel.refreshTable();
-    }
-
-    /**
      * Saves a network design to a {@code .n2p} file.
      *
      * @since 0.3.0
@@ -538,6 +423,6 @@ public class TopologyPanel extends JPanel implements ActionListener//FrequentisB
 
     public final void updateMultilayerPanel()
     {
-        multilayerControlPanel.refreshTable();
+        sideBar.getMultilayerControlPanel().refreshTable();
     }
 }
